# further optimize selectByKind for special case of all fitting servers